refactor(web): move blog queries to lib/data/
