+++ apps/web/src/app/(dashboard)/(home)/page.tsx
@@ -9,7 +9,7 @@ import { StructuredData } from "@web/components/structured-data";
 import { TopMakesByYear } from "@web/components/top-makes-by-year";
 import { TotalNewCarRegistrationsByYear } from "@web/components/total-new-car-registrations-by-year";
 import { SITE_TITLE, SITE_URL } from "@web/config";
-import { getAllPosts } from "@web/utils/blog-queries";
+import { getAllPosts } from "@web/lib/data/posts";
 import { getLatestCOEResults } from "@web/utils/cached-api";
 import type { WebSite, WithContext } from "schema-dts";
